Fix cross C/C++ linkage
Imagine if the header was included in a C file while
CLOWNRESAMPLER_IMPLEMENTATION was defined, but the header was also
included in a C++ file that used its functions. Prior to this commit,
there would be linker errors as it would try to resolve C++ versions
of C function names.