Make the user handle outputting the samples.
Now the user can decide whether the samples need clamping and
truncating to 16-bit, or whether they should be copied or mixed into
the output buffer, or whatever. My Mega Drive emulator uses this to
upmix the PSG samples to stereo and mix them with the FM samples
during resampling, to avoid the use of an intermediary buffer and an
additional separate audio processing step.